+++ components/button/src/Button.tsx
@@ -84,6 +84,13 @@ export interface ButtonProps extends Omit<ButtonHTMLAttributes<HTMLButtonElement
 	iconColor?: string;
 
 	/**
+	 * Icon size
+	 *
+	 * Defaults to button size prop
+	 */
+	iconSize?: 'small' | 'medium' | 'large' | 'xlarge' | ('small' | 'medium' | 'large' | 'xlarge')[];
+
+	/**
 	 * Justify align button children
 	 */
 	justify?: boolean;
@@ -117,12 +124,6 @@ export interface ButtonProps extends Omit<ButtonHTMLAttributes<HTMLButtonElement
 	 * horizontalPadding for link style
 	 */
 	horizontalPadding?: boolean;
-	/**
-
-	/**
-	 * min-width
-	 */
-	minWidth?: string;
 	/**
 
 	 * The override API
@@ -161,6 +162,7 @@ export const Button = forwardRef(
 			iconAfter,
 			iconBefore,
 			iconColor,
+			iconSize,
 			assistiveText,
 			type,
 			dropdown,
@@ -175,7 +177,6 @@ export const Button = forwardRef(
 			justify = false,
 			disabled = false,
 			horizontalPadding = true,
-			minWidth = 'auto',
 			...rest
 		}: ButtonProps,
 		ref
@@ -204,6 +205,7 @@ export const Button = forwardRef(
 			iconAfter,
 			iconBefore,
 			iconColor,
+			iconSize,
 			justify,
 			disabled,
 			assistiveText,
@@ -212,7 +214,6 @@ export const Button = forwardRef(
 			dropdown,
 			onClick,
 			horizontalPadding,
-			minWidth,
 			overrides: componentOverrides,
 			...rest,
 		};
@@ -241,6 +242,7 @@ export const Button = forwardRef(
 							iconBefore={iconBefore}
 							dropdown={dropdown}
 							iconColor={iconColor}
+							iconSize={iconSize}
 						>
 							{children}
 						</Content>
@@ -307,6 +309,15 @@ Button.propTypes = {
 	 */
 	iconColor: PropTypes.string,
 	/**
+	 * Icon size
+	 *
+	 * Defaults to button size prop
+	 */
+	iconSize: PropTypes.oneOfType([
+		PropTypes.oneOf(['large', 'medium', 'small', 'xlarge']),
+		PropTypes.arrayOf(PropTypes.oneOf(['large', 'medium', 'small', 'xlarge'])),
+	]),
+	/**
 	 * Justify align button children
 	 */
 	justify: PropTypes.bool,
@@ -314,11 +325,6 @@ Button.propTypes = {
 	 * Button look
 	 */
 	look: PropTypes.oneOf(['faint', 'hero', 'link', 'primary', 'unstyled']),
-	/**
-	 * /**
-	 * min-width
-	 */
-	minWidth: PropTypes.string,
 	/**
 	 * Handler to be called on click
 	 */