+++ components/text-input/tests/unit/text-input.spec.js
@@ -84,7 +84,6 @@ describe('TextInput component', () => {
 		const inputNode = screen.getByDisplayValue('sample');
 		expect(inputNode).toBeInTheDocument();
 		expect(inputNode.value).toEqual('sample');
-		expect(inputNode.className).toEqual('css-4t3uc7-textInput');
 		const user = userEvent.setup();
 		await user.tripleClick(inputNode);
 		await user.keyboard('{backspace}');