+++ helpers/example/_utils.js
@@ -10,7 +10,11 @@ const fs = require('fs');
  */
 const slugFromFilename = (filename) => {
 	if (filename.match(/^[0-9][0-9]-/)) {
-		return filename.slice(3, -3);
+		if (filename.match(/(.js|.ts)$/)) {
+			return filename.slice(3, -3);
+		} else if (filename.match(/.tsx$/)) {
+			return filename.slice(3, -4);
+		}
 	} else if (filename.match(/(.js|.ts)$/)) {
 		return filename.slice(0, -3);
 	} else if (filename.match(/.tsx$/)) {